module-template/backend/requirements.txt
# MODULE_DISPLAY_NAME Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
python-multipart>=0.0.6

# Database (optional - uncomment if needed)
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0
# alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.0
